Deleted closestFood.py, moved _ClosestFoodSearchProblem to featureExtractors.py, corrected style.
