+++ pacai/agents/search/base.py
@@ -50,7 +50,7 @@ class SearchAgent(BaseAgent):
         # Fetch the heuristic.
         heuristic = reflection.qualifiedImport(heuristicName)
         logging.info('[SearchAgent] using function %s and heuristic %s.' %
-            (functionName, heuristicName))
+                (functionName, heuristicName))
 
         # Bind the heuristic.
         return lambda x: function(x, heuristic = heuristic)
