+++ pacai/core/featureExtractors.py
@@ -1,11 +1,59 @@
-# Feature extractors for Pacman game states.
+# Feature extractors for Pacman game states and a private search problem to find
+# the closest food.
 
 import abc
 
 from pacai.core import game
-from pacai.core.search import _ClosestFoodSearchProblem
+from pacai.core.game import Actions
+from pacai.core.game import Directions
+from pacai.core.search.problem import SearchProblem
 from pacai.student import search
 from pacai.util import counter
+
+class _ClosestFoodSearchProblem(SearchProblem):
+
+    """
+    A private search problem associated with finding the closest food
+    from an initial position on the map.
+
+    Search State: a tuple (x,y) representing the current position in the
+    search
+    """
+
+    # Search problem requires the initial position to search from, the food
+    # and walls on the map
+    def __init__(self, initPos, food, walls):
+        self._start = initPos
+        self._foodGrid = food
+        self._walls = walls
+
+    def startingState(self):
+        return self._start
+
+    # Goal state is where the current position is at the same location as food
+    def isGoal(self, state):
+        return self._foodGrid[state[0]][state[1]]
+
+    def successorStates(self, state):
+        successors = []
+        for direction in [Directions.NORTH, Directions.SOUTH, Directions.EAST, Directions.WEST]:
+            x, y = state
+            dx, dy = Actions.directionToVector(direction)
+            nextx, nexty = int(x + dx) + int(y + dy)
+            if not self._walls[nextx][nexty]:
+                successors.append((nextx, nexty))
+        return successors
+
+    def actionsCost(self, actions):
+        x, y = self._start
+        cost = 0
+        for action in actions:
+            dx, dy = Actions.directionToVector(action)
+            x, y = int(x + dx) + int(y + dy)
+            if self._walls[x][y]:
+                return 999999
+            cost += 1
+        return cost
 
 class FeatureExtractor(abc.ABC):
     @abc.abstractmethod
@@ -32,8 +80,6 @@ def _closestFood(pos, food, walls):
     """
 
     x1, y1 = pos
-
-    assert not walls[x1][y1], 'pos is a wall: ' + pos
 
     prob = _ClosestFoodSearchProblem(pos, food, walls)
     return len(search.bfs(prob))
